No import in vanilla js
